transfer most of functions
