+++ include/scaluq/gate/param_gate_probablistic.hpp
@@ -17,22 +17,7 @@ class ParamProbablisticGateImpl : public ParamGateBase<Fp> {
 
 public:
     ParamProbablisticGateImpl(const std::vector<Fp>& distribution,
-                              const std::vector<std::variant<Gate<Fp>, ParamGate<Fp>>>& gate_list)
-        : ParamGateBase<Fp>(0, 0), _distribution(distribution), _gate_list(gate_list) {
-        std::uint64_t n = distribution.size();
-        if (n == 0) {
-            throw std::runtime_error("At least one gate is required.");
-        }
-        if (n != gate_list.size()) {
-            throw std::runtime_error("distribution and gate_list have different size.");
-        }
-        _cumlative_distribution.resize(n + 1);
-        std::partial_sum(
-            distribution.begin(), distribution.end(), _cumlative_distribution.begin() + 1);
-        if (std::abs(_cumlative_distribution.back() - 1.) > 1e-6) {
-            throw std::runtime_error("Sum of distribution must be equal to 1.");
-        }
-    }
+                              const std::vector<std::variant<Gate<Fp>, ParamGate<Fp>>>& gate_list);
     const std::vector<EitherGate>& gate_list() const { return _gate_list; }
     const std::vector<Fp>& distribution() const { return _distribution; }
 
@@ -67,51 +52,16 @@ public:
             "ParamProbablisticGateImpl.");
     }
 
-    std::shared_ptr<const ParamGateBase<Fp>> get_inverse() const override {
-        std::vector<EitherGate> inv_gate_list;
-        inv_gate_list.reserve(_gate_list.size());
-        std::ranges::transform(
-            _gate_list, std::back_inserter(inv_gate_list), [](const EitherGate& gate) {
-                return std::visit([](const auto& g) { return EitherGate{g->get_inverse()}; }, gate);
-            });
-        return std::make_shared<const ParamProbablisticGateImpl>(_distribution, inv_gate_list);
-    }
+    std::shared_ptr<const ParamGateBase<Fp>> get_inverse() const override;
     internal::ComplexMatrix<Fp> get_matrix(Fp) const override {
         throw std::runtime_error(
             "ParamProbablisticGateImpl::get_matrix(): This function must not be used in "
             "ParamProbablisticGateImpl.");
     }
 
-    void update_quantum_state(StateVector<Fp>& state_vector, Fp param) const override {
-        Random random;
-        Fp r = random.uniform();
-        std::uint64_t i = std::distance(_cumlative_distribution.begin(),
-                                        std::ranges::upper_bound(_cumlative_distribution, r)) -
-                          1;
-        if (i >= _gate_list.size()) i = _gate_list.size() - 1;
-        const auto& gate = _gate_list[i];
-        if (gate.index() == 0) {
-            std::get<0>(gate)->update_quantum_state(state_vector);
-        } else {
-            std::get<1>(gate)->update_quantum_state(state_vector, param);
-        }
-    }
+    void update_quantum_state(StateVector<Fp>& state_vector, Fp param) const override;
 
-    std::string to_string(const std::string& indent) const override {
-        std::ostringstream ss;
-        const auto dist = distribution();
-        ss << indent << "Gate Type: Probablistic\n";
-        for (std::size_t i = 0; i < dist.size(); ++i) {
-            ss << indent << "  --------------------\n";
-            ss << indent << "  Probability: " << dist[i] << "\n";
-            std::visit(
-                [&](auto&& arg) {
-                    ss << arg->to_string(indent + "  ") << (i == dist.size() - 1 ? "" : "\n");
-                },
-                gate_list()[i]);
-        }
-        return ss.str();
-    }
+    std::string to_string(const std::string& indent) const override;
 };
 }  // namespace internal
