+++ src/CMakeLists.txt
@@ -1,7 +1,19 @@
 cmake_minimum_required(VERSION 3.21)
 
 target_sources(scaluq PRIVATE
+    gate/gate_matrix.cpp
+    gate/gate_pauli.cpp
+    gate/gate_probablistic.cpp
+    gate/gate_standard.cpp
     gate/gate.cpp
+    gate/param_gate.cpp
+    gate/param_gate_pauli.cpp
+    gate/param_gate_probablistic.cpp
+    gate/param_gate_standard.cpp
+    gate/param_gate.cpp
+    gate/update_ops_dense_matrix.cpp
+    gate/update_ops_sparse_matrix.cpp
+    gate/update_ops_standard.cpp
     state/state_vector.cpp
     state/state_vector_batched.cpp
     operator/apply_pauli.cpp
