+++ tests/CMakeLists.txt
@@ -14,11 +14,10 @@ add_executable(scaluq_test EXCLUDE_FROM_ALL
     state/state_vector_batched_test.cpp
 )
 
-target_link_libraries(scaluq_test PRIVATE
+target_link_libraries(scaluq_test PUBLIC 
     scaluq
     Kokkos::kokkos
     GTest::gtest_main
-    Eigen3::Eigen
 )
-
-target_include_directories(scaluq_test PRIVATE ${PROJECT_SOURCE_DIR}/scaluq)
+target_include_directories(scaluq_test PRIVATE $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include> $<INSTALL_INTERFACE:include>)
+target_include_directories(scaluq_test PRIVATE ${eigen_SOURCE_DIR})
