+++ tests/circuit/param_circuit_test.cpp
@@ -1,14 +1,9 @@
 #include <gtest/gtest.h>
 
-#include <circuit/circuit.hpp>
-#include <gate/gate_factory.hpp>
-#include <gate/param_gate_factory.hpp>
-#include <numbers>
-#include <state/state_vector.hpp>
-#include <types.hpp>
-#include <util/random.hpp>
+#include <scaluq/circuit/circuit.hpp>
+#include <scaluq/gate/gate_factory.hpp>
+#include <scaluq/gate/param_gate_factory.hpp>
 
-#include "../test_environment.hpp"
 #include "../util/util.hpp"
 
 using namespace scaluq;
