+++ tests/state/state_vector_batched_test.cpp
@@ -1,9 +1,6 @@
 #include <gtest/gtest.h>
 
-#include <Eigen/Core>
-#include <cmath>
-#include <state/state_vector_batched.hpp>
-#include <util/utility.hpp>
+#include <scaluq/state/state_vector_batched.hpp>
 
 #include "../test_environment.hpp"
 #include "../util/util.hpp"
